refactored and improved TC architecture
