+++ src/tmtccmd/pus_tc/service_17_test.py
@@ -2,6 +2,7 @@ import enum
 
 from tmtccmd.core.definitions import QueueCommands
 from tmtccmd.pus_tc.base import PusTelecommand, TcQueueT
+from tmtccmd.config.globals import get_global_apid
 
 
 class Srv17Subservices(enum.IntEnum):
@@ -9,11 +10,15 @@ class Srv17Subservices(enum.IntEnum):
     GEN_EVENT = 128
 
 
-def pack_service17_ping_command(ssc: int) -> PusTelecommand:
-    return PusTelecommand(service=17, subservice=Srv17Subservices.PING_CMD, ssc=ssc)
+def pack_service17_ping_command(ssc: int, apid: int = -1) -> PusTelecommand:
+    if apid == -1:
+        apid = get_global_apid()
+    return PusTelecommand(service=17, subservice=Srv17Subservices.PING_CMD, ssc=ssc, apid=apid)
 
 
-def pack_generic_service17_test(init_ssc: int, tc_queue: TcQueueT) -> int:
+def pack_generic_service17_test(init_ssc: int, tc_queue: TcQueueT, apid: int = -1) -> int:
+    if apid == -1:
+        apid = get_global_apid()
     new_ssc = init_ssc
     tc_queue.appendleft((QueueCommands.PRINT, "Testing Service 17"))
     # ping test
@@ -22,17 +27,19 @@ def pack_generic_service17_test(init_ssc: int, tc_queue: TcQueueT) -> int:
     new_ssc += 1
     # enable event
     tc_queue.appendleft((QueueCommands.PRINT, "Testing Service 17: Enable Event"))
-    command = PusTelecommand(service=5, subservice=5, ssc=new_ssc)
+    command = PusTelecommand(service=5, subservice=5, ssc=new_ssc, apid=apid)
     tc_queue.appendleft(command.pack_command_tuple())
     new_ssc += 1
     # test event
     tc_queue.appendleft((QueueCommands.PRINT, "Testing Service 17: Trigger event"))
-    command = PusTelecommand(service=17, subservice=Srv17Subservices.GEN_EVENT, ssc=new_ssc)
+    command = PusTelecommand(
+        service=17, subservice=Srv17Subservices.GEN_EVENT, ssc=new_ssc, apid=apid
+    )
     tc_queue.appendleft(command.pack_command_tuple())
     new_ssc += 1
     # invalid subservice
     tc_queue.appendleft((QueueCommands.PRINT, "Testing Service 17: Invalid subservice"))
-    command = PusTelecommand(service=17, subservice=243, ssc=new_ssc)
+    command = PusTelecommand(service=17, subservice=243, ssc=new_ssc, apid=apid)
     tc_queue.appendleft(command.pack_command_tuple())
     new_ssc += 1
     return new_ssc
